Split tests into different categories
