+++ modules/mtl/input/src/main/scala-2/fix/MTLSubmarineWrappedRaiseTest.scala
@@ -1,0 +1,22 @@
+/*
+rule = TypelevelMTLSubmarine
+ */
+package fix
+
+import cats.mtl._
+import cats.effect._
+import cats.syntax.all._
+
+
+// scalafmt: { maxColumn = 160 }
+object WrappedRaise {
+
+  trait Wrapper[A]
+
+  def normal[F[_]: Async](implicit wrapper: Wrapper[Raise[F, String]]): F[Unit] =
+    Async[F].unit
+
+  def handlingNormal[F[_]: Async](implicit wrapper: Wrapper[Raise[F, String]]): F[Unit] =
+    normal[F].attempt.void
+
+}
